stats errors diagnosing with askbcs
