+++ client/src/App.js
@@ -6,8 +6,9 @@ import UserLogin from "./pages/UserLogin";
 import UserSignup from "./pages/UserSignup";
 import User2Profile from "./pages/User2Profile";
 import PastEvents from "./pages/PastEvents";
-import './App.css';
 import Stats from "./pages/Stats";
+import './App.css';
+import * as ReactBootStrap from "react-bootstrap";
 
 function App() {
   return (
@@ -15,7 +16,7 @@ function App() {
       <div>
         <Switch>
           <Route exact path="/" component={Home}>
-            {/* <Home /> */}
+            <Home />
           </Route>
           <Route exact path={["/searchresult"]}>
             <SearchResults />
