+++ package.json
@@ -27,6 +27,8 @@
     "@material-ui/icons": "^4.11.2",
     "axios": "^0.18.0",
     "bcryptjs": "^2.4.3",
+    "bootstrap": "^4.6.0",
+    "chart.js": "^3.1.1",
     "connect-mongo": "^4.4.1",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
@@ -40,6 +42,8 @@
     "morgan": "^1.10.0",
     "node": "^15.12.0",
     "passport": "^0.4.1",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "react-bootstrap": "^1.5.2",
+    "react-chartjs-2": "^2.11.1"
   }
 }
