+++ routes/api/pastevents.js
@@ -3,6 +3,7 @@ const pastEventController = require("../../controllers/pastEventController");
 
 router.route("/")
   .get(pastEventController.findAll)
+  // .get(pastEventController.getArtistTotal)
   .post(pastEventController.create);
 
 router
